+++ src/engine/random.rs
@@ -39,6 +39,6 @@ impl Engine for RandomEngine {
     #[allow(unused_variables)]
     fn gen_move(&self, color: Color, game: &Game) -> Move {
         let moves = game.legal_moves();
-        moves[random::<uint>() % moves.len()]
+        moves[random::<usize>() % moves.len()]
     }
 }
